Refactor: Extract test utility function for reuse
Tried exporting it from render/utils/__tests__/animation-state.test.ts,
but that would implicitly also run the tests within...
